requirements.txt
datetime-utils @ git+https://github.com/reimibeta/django-datetime-utils.git
Django==3.1.7
django-admin-list-filter-dropdown==1.0.3
django-cleanup==5.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-flex-fields==0.8.9
html-render-utils @ git+https://github.com/reimibeta/django-html-render-utils.git
image-utils @ git+https://github.com/reimibeta/django-image-utils.git
mysqlclient==2.0.3
Pillow==8.2.0
rest-framework-utils @ git+https://github.com/reimibeta/django-rest-framework-utils.git
stock-algorithms @ git+https://github.com/reimibeta/django-stock-utils.git
user-models @ git+https://github.com/reimibeta/django-user-models.git
wallet-models @ git+https://github.com/reimibeta/django-wallet-models.git
